// src/types/gamesTypes.ts
import type { Dispatch } from "react";

export type OptionsType = "paper" | "rock" | "scissors";

export type GameStateType = {
  points: number;
  playerChoice: null | OptionsType;
  computerChoice: null | OptionsType;
  step: number;
  finalResult: string;
};

export type GameActionType =
  | {
      type: "playerChoice/add";
      payload: OptionsType;
    }
  | { type: "computerChoice/add"; payload: OptionsType }
  | { type: "reset" }
  | { type: "step/update"; payload: number }
  | { type: "over" }
  | { type: "update/score"; payload: number };

export type GameContextType = {
  state: GameStateType;
  dispatch: Dispatch<GameActionType>;
};